examples/fastapi_testing_app/requirements.txt
-e ../..
python-dotenv
fastapi
uvicorn
openai
tiktoken
langchain-openai
google-cloud-speech
google-cloud-texttospeech
google-cloud-aiplatform
langchain-google-genai


